+++ addOns/client/src/main/java/org/zaproxy/addon/client/spider/PopupMenuSpider.java
@@ -22,6 +22,7 @@ package org.zaproxy.addon.client.spider;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.model.SiteNode;
 import org.zaproxy.addon.client.ExtensionClientIntegration;
+import org.zaproxy.addon.commonlib.MenuWeights;
 import org.zaproxy.zap.view.messagecontainer.http.HttpMessageContainer;
 import org.zaproxy.zap.view.popup.PopupMenuItemSiteNodeContainer;
 
@@ -50,6 +51,14 @@ public class PopupMenuSpider extends PopupMenuItemSiteNodeContainer {
     @Override
     public int getParentMenuIndex() {
         return ATTACK_MENU_INDEX;
+    }
+
+    public int getParentWeight() {
+        return MenuWeights.MENU_ATTACK_WEIGHT;
+    }
+
+    public int getWeight() {
+        return MenuWeights.MENU_ATTACK_CLIENT_WEIGHT;
     }
 
     @Override