+++ addOns/spider/src/main/java/org/zaproxy/addon/spider/PopupMenuItemSpiderDialog.java
@@ -21,6 +21,7 @@ package org.zaproxy.addon.spider;
 
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.model.SiteNode;
+import org.zaproxy.addon.commonlib.MenuWeights;
 import org.zaproxy.zap.view.messagecontainer.http.HttpMessageContainer;
 import org.zaproxy.zap.view.popup.PopupMenuItemSiteNodeContainer;
 
@@ -79,5 +80,13 @@ public class PopupMenuItemSpiderDialog extends PopupMenuItemSiteNodeContainer {
             default:
                 return true;
         }
+    }
+
+    public int getParentWeight() {
+        return MenuWeights.MENU_ATTACK_WEIGHT;
+    }
+
+    public int getWeight() {
+        return MenuWeights.MENU_ATTACK_SPIDER_WEIGHT;
     }
 }