+++ addOns/websocket/src/main/java/org/zaproxy/zap/extension/websocket/ui/PopupIncludeWebSocketContextMenu.java
@@ -22,6 +22,7 @@ package org.zaproxy.zap.extension.websocket.ui;
 import java.awt.Component;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.extension.ExtensionPopupMenuItem;
+import org.zaproxy.addon.commonlib.MenuWeights;
 import org.zaproxy.zap.model.Context;
 import org.zaproxy.zap.view.popup.PopupMenuItemContextInclude;
 
@@ -35,6 +36,10 @@ public class PopupIncludeWebSocketContextMenu extends PopupMenuItemContextInclud
     @Override
     public String getParentMenuName() {
         return MENU_NAME;
+    }
+
+    public int getParentWeight() {
+        return MenuWeights.MENU_INC_CHANNEL_CONTEXT_WEIGHT;
     }
 
     @Override