+++ addOns/requester/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [7.4.0] - 2023-12-8
+
 ## Added 
 - Added Button to LowerCase Headers Name