+++ addOns/requester/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [7.5.0] - 2024-03-25
 ### Added 
 - Button to lowercase request header names (Issue 8176).
 - Support for menu weights (Issue 8369)
@@ -99,6 +99,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[7.5.0]: https://github.com/zaproxy/zap-extensions/releases/requester-v7.5.0
 [7.4.0]: https://github.com/zaproxy/zap-extensions/releases/requester-v7.4.0
 [7.3.0]: https://github.com/zaproxy/zap-extensions/releases/requester-v7.3.0
 [7.2.0]: https://github.com/zaproxy/zap-extensions/releases/requester-v7.2.0